Refactor out of distribution models and utils
